% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ieITH.R
\name{ieITH}
\alias{ieITH}
\title{ieITH}
\usage{
ieITH(data_tumor, data_normal = NULL, data_type)
}
\arguments{
\item{data_tumor}{A dataframe or matrix containing one of the multi-omics profiles in tumors: 
1) For DNA methylation, mRNA expression, miRNA expression, lncRNA expression, and protein expressionin, the row name is probes or RNAs or proteins and the column name is tumor sample ID;
2) For "maf" files, it at least includes two columns ("sample" and "vaf");
3) For "SNP6" files,it at least includes two columns ("sample" and "value").}

\item{data_normal}{A dataframe or matrix containing matched normal sample profiles for DNA methylation, mRNA expression, miRNA expression, lncRNA expression, and protein expressionin, 
the row names of data_normal should be matched with the row names of data_tumor.If the data_normal is NULL, calculate the IE-based ITH score by with tumor samples.}

\item{data_type}{A character belonging to one of the c("cnv", "mut", "met", "lnc", "mir", "mrn", "pro").}
}
\value{
A dataframe with 3 columns:
\item{sample}{Tumor samples to be calculated.}
\item{ieITH_score}{The ieITH score of each sample.}
\item{score_type}{The type of IE-based ITH score.}
}
\description{
\code{ieITH} ieITH evaluates the ITH level of each bulk tumor sample based on one of the multi-omics profiles in tumors.
}
\details{
The multi-omics profiles could be "SNP6" files, "maf" files, DNA methylation, mRNA expression, miRNA expression, lncRNA expression, and protein expression.
}
\examples{
#example1
path1 <- system.file("extdata", "example_mrnITH_tumor.txt", package = "ieITH", mustWork = TRUE)
path2 <- system.file("extdata", "example_mrnITH_normal.txt", package = "ieITH", mustWork = TRUE)
data_tumor <- read.table(path1, stringsAsFactors = FALSE, header = TRUE, check.names = FALSE, sep = "\t", quote = "", row.names = 1)
data_normal <- read.table(path2, stringsAsFactors = FALSE, header = TRUE, check.names = FALSE, sep = "\t", quote = "", row.names = 1)
mrnITH <- ieITH(data_tumor, data_normal, "mrn")
#example2
path <- system.file("extdata", "example_cnvITH.txt", package = "ieITH", mustWork = TRUE)
data_tumor <- read.table(path, stringsAsFactors = FALSE, header = TRUE, check.names = FALSE, sep = "\t", quote = "")
cnvITH <- ieITH(data_tumor, data_normal = NULL, "cnv")
#example3
path <- system.file("extdata", "example_mutITH.txt", package = "ieITH", mustWork = TRUE)
data_tumor <- read.table(path, stringsAsFactors = FALSE, header = TRUE, check.names = FALSE, sep = "\t", quote = "")
mutITH <- ieITH(data_tumor, data_normal = NULL, "mut")
}
