% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intITH.R
\name{intITH}
\alias{intITH}
\title{intITH}
\usage{
intITH(input_data_int)
}
\arguments{
\item{input_data_int}{A dataframe containing 3 columns: "sample", "ieITH_score" and "score_type".}
}
\value{
A dataframe with 2 columns:
\item{sample}{Tumor samples to be calculated.}
\item{intITH_score}{The intITH score of each sample.}
}
\description{
\code{intITH} intITH is the average of different IE-based ITH scores after scaling into [0-1] within each type of ITH scores.
}
\examples{
path <- system.file("extdata", "example_intITH.txt", package = "ieITH", mustWork = TRUE)
input_data_int <- read.table(path, stringsAsFactors = FALSE, header = TRUE, check.names = FALSE, sep = "\t", quote = "")
ieITH_score <- intITH(input_data_int)
}
